refactor https parsing
