+++ packages/core/src/urlAdapters.ts
@@ -1,3 +1,21 @@
+export interface UrlAdapter {
+    contentType?: "text/plain" | "application/json"
+
+    /**
+     * Given a friendly URL, return a URL that can be used to fetch the content.
+     * @param url
+     * @returns
+     */
+    matcher: (url: string) => string
+
+    /**
+     * Convers the body of the response to a string.
+     * @param body
+     * @returns
+     */
+    adapter?: (body: string | any) => string | undefined
+}
+
 export const defaultUrlAdapters: UrlAdapter[] = [
     {
         matcher: (url) => {
